+++ examples/gunicorn_conf.py
@@ -1,12 +1,9 @@
-import logging
+"""Gunicorn configuration for ``examples.gunicorn_runner:app``."""
 
 from request_id_helper import RequestIdFormatter
 
-# Gunicorn config
 workers = 1
 worker_class = "uvicorn.workers.UvicornWorker"
-
-# Logging Options
 loglevel = "info"
 logconfig_dict = {
     "version": 1,
@@ -14,8 +11,7 @@ logconfig_dict = {
     "formatters": {
         "default": {
             "()": RequestIdFormatter,
-            "format": "[%(asctime)s] %(levelname)s [%(request_id)s] %(name)s | %(message)s",
-            "datefmt": "%d/%b/%Y %H:%M:%S",
+            "format": "%(levelname)s [%(request_id)s] %(name)s: %(message)s",
         },
     },
     "handlers": {
@@ -25,15 +21,7 @@ logconfig_dict = {
         },
     },
     "loggers": {
-        "gunicorn.error": {
-            "level": "INFO",
-            "handlers": ["console"],
-            "propagate": False,
-        },
-        "gunicorn.access": {
-            "level": "INFO",
-            "handlers": ["console"],
-            "propagate": False,
-        },
+        "gunicorn.error": {"handlers": ["console"], "level": "INFO", "propagate": False},
+        "gunicorn.access": {"handlers": ["console"], "level": "INFO", "propagate": False},
     },
 }