+++ src/starlette_request_id/__init__.py
@@ -1,7 +1,11 @@
+"""Request-ID context and Starlette middleware."""
+
 from request_id_helper import LogExtraFactory, init_logger, request_id_ctx
 
 from .constants import REQUEST_ID_HEADER
 from .middleware import RequestIdMiddleware
+
+__version__ = "2.0.0"
 
 __all__ = [
     "REQUEST_ID_HEADER",
@@ -9,4 +13,5 @@ __all__ = [
     "LogExtraFactory",
     "request_id_ctx",
     "init_logger",
+    "__version__",
 ]