+++ src/starlette_request_id/constants.py
@@ -1,0 +1,3 @@
+"""Public constants for request-ID propagation."""
+
+REQUEST_ID_HEADER = "x-request-id"