+++ starlette_request_id/constants.py
@@ -1,1 +1,0 @@
-REQUEST_ID_HEADER = "x-request-id"